pyproject.toml
[project]
name = "mcproxy"
version = "3.4.0"
description = "Lightweight MCP gateway that aggregates multiple stdio MCP servers through namespaced SSE endpoints"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Bkuri", email = "bk@kuri.casa" }
]
keywords = ["mcp", "gateway", "proxy", "sse", "stdio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-json-logger>=2.0.7",
    "fastmcp>=0.1.0",
    "orjson>=3.9.0",
    "cryptography>=42.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mcproxy = "main:main"

[project.urls]
Homepage = "https://github.com/bkuri/mcproxy"
Documentation = "https://github.com/bkuri/mcproxy#readme"
Repository = "https://github.com/bkuri/mcproxy"
Issues = "https://github.com/bkuri/mcproxy/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
